requirements.txt
chatterbot2
chatterbot-corpus